requirements.txt
streamlit
google-generativeai
pdfplumber
python-docx
python-dotenv
textblob
googletrans